Generify DimensionProperty
