+++ src/main/java/gregtech/api/recipes/RecipeBuilder.java
@@ -13,6 +13,7 @@ import gregtech.api.recipes.ingredients.*;
 import gregtech.api.recipes.ingredients.nbtmatch.NBTCondition;
 import gregtech.api.recipes.ingredients.nbtmatch.NBTMatcher;
 import gregtech.api.recipes.recipeproperties.CleanroomProperty;
+import gregtech.api.recipes.recipeproperties.DimensionProperty;
 import gregtech.api.recipes.recipeproperties.IRecipePropertyStorage;
 import gregtech.api.recipes.recipeproperties.RecipeProperty;
 import gregtech.api.recipes.recipeproperties.RecipePropertyStorage;
@@ -39,6 +40,9 @@ import com.cleanroommc.groovyscript.api.GroovyLog;
 import com.cleanroommc.groovyscript.api.IIngredient;
 import com.cleanroommc.groovyscript.helper.ingredient.OreDictIngredient;
 import crafttweaker.CraftTweakerAPI;
+import it.unimi.dsi.fastutil.ints.IntArrayList;
+import it.unimi.dsi.fastutil.ints.IntList;
+import it.unimi.dsi.fastutil.ints.IntLists;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -137,8 +141,43 @@ public class RecipeBuilder<R extends RecipeBuilder<R>> {
         return (R) this;
     }
 
+    public R dimension(int dimensionID) {
+        IntList dimensionIDs = getDimensionIDs();
+        if (dimensionIDs == IntLists.EMPTY_LIST) {
+            dimensionIDs = new IntArrayList();
+            this.applyProperty(DimensionProperty.getInstance(), dimensionIDs);
+        }
+        dimensionIDs.add(dimensionID);
+        return (R) this;
+    }
+
+    public IntList getDimensionIDs() {
+        return this.recipePropertyStorage == null ? IntLists.EMPTY_LIST :
+                this.recipePropertyStorage.getRecipePropertyValue(DimensionProperty.getInstance(),
+                        IntLists.EMPTY_LIST);
+    }
+
     public boolean applyProperty(@NotNull String key, @Nullable Object value) {
-        if (key.equals(CleanroomProperty.KEY)) {
+        if (key.equals(DimensionProperty.KEY)) {
+            if (value instanceof Integer) {
+                this.dimension((Integer) value);
+            } else if (value instanceof List && !((List<?>) value).isEmpty() &&
+                    ((List<?>) value).get(0) instanceof Integer) {
+                        IntList dimensionIDs = getDimensionIDs();
+                        if (dimensionIDs == IntLists.EMPTY_LIST) {
+                            dimensionIDs = new IntArrayList();
+                            this.applyProperty(DimensionProperty.getInstance(), dimensionIDs);
+                        }
+                        dimensionIDs.addAll((List<Integer>) value);
+                    } else {
+                        if (isCTRecipe) {
+                            CraftTweakerAPI.logError("Dimension for Dimension Property needs to be an Integer");
+                            return false;
+                        }
+                        throw new IllegalArgumentException("Invalid Dimension Property Type!");
+                    }
+            return true;
+        } else if (key.equals(CleanroomProperty.KEY)) {
             if (value instanceof CleanroomType) {
                 this.cleanroom((CleanroomType) value);
             } else if (value instanceof String) {
@@ -976,6 +1015,7 @@ public class RecipeBuilder<R extends RecipeBuilder<R>> {
                 .append("EUt", EUt)
                 .append("hidden", hidden)
                 .append("cleanroom", getCleanroom())
+                .append(DimensionProperty.getInstance().getKey(), getDimensionIDs().toString())
                 .append("recipeStatus", recipeStatus)
                 .toString();
     }
