+++ src/main/java/gregtech/common/metatileentities/electric/MetaTileEntityGasCollector.java
@@ -8,7 +8,7 @@ import gregtech.api.metatileentity.interfaces.IGregTechTileEntity;
 import gregtech.api.recipes.Recipe;
 import gregtech.api.recipes.RecipeMap;
 import gregtech.api.recipes.RecipeMaps;
-import gregtech.api.recipes.recipeproperties.GasCollectorDimensionProperty;
+import gregtech.api.recipes.recipeproperties.DimensionProperty;
 import gregtech.client.renderer.ICubeRenderer;
 import gregtech.client.renderer.texture.Textures;
 
@@ -40,7 +40,7 @@ public class MetaTileEntityGasCollector extends SimpleMachineMetaTileEntity {
     }
 
     protected boolean checkRecipe(@NotNull Recipe recipe) {
-        for (int dimension : recipe.getProperty(GasCollectorDimensionProperty.getInstance(), IntLists.EMPTY_LIST)) {
+        for (int dimension : recipe.getProperty(DimensionProperty.getInstance(), IntLists.EMPTY_LIST)) {
             if (dimension == this.getWorld().provider.getDimension()) {
                 return true;
             }
